feat. 고객 조회 반환 값 로직 추가 ( remainingTime, remainingPeriod, usedTime, lateCount, absenceCount )

+++ src/main/java/com/erp/erp/domain/plan/common/entity/Plan.java
@@ -33,7 +33,7 @@ public class Plan {
   private int price;
 
   @PositiveOrZero
-  private int availableTime;
+  private double availableTime;
 
   @NotNull
   @PositiveOrZero
@@ -41,7 +41,7 @@ public class Plan {
 
   @Builder
   public Plan(PlanType planType, LicenseType licenseType, CourseType courseType, String name,
-      int price, int availableTime, int availablePeriod) {
+      int price, double availableTime, int availablePeriod) {
     this.planType = planType;
     this.licenseType = licenseType;
     this.courseType = courseType;
